About to refactor to new pkg neilotoole/fifomu
